remove cwebmac.tex and cwebmal.tex
First cwebmac.tex (with cwebmac-fix.patch) was used because cwebmal.tex was not available in post-install.sh (it is in TEXMFHOME/tex/).
Then we copied cwebmal.tex to cweb-git repo because we needed it in docs. Now this became possible
to use cwebmal.tex with cw. But then we moved cweb-git out of installer - so the need for cwebmal.tex also disappeared.